works with CDB  0.9.0a747640
for unfound fields, needed to return {<<"">>,<<"">>} in
find_all_content/2 and find_all_fields/2 ; returning []
used to work (though that was wrong).  Not going to use
the atoms version, though, because saving a bit of typing
for the one-time map fun isn't worth the penalty to me
of converting back to <<"list">> for each of possibly
millions of documents.